wip new user managements

+++ routes/fileshare/create-repos.js
@@ -1,29 +1,38 @@
-const {require_connection, get_user_private_data} = require("../../src/session_utils");
+const {require_connection, get_user_private_data, session_data, error_403, public_data, events} = require("../../src/session_utils");
 const Repos = require('../../src/database/tables/repos')
 const Users = require("../../src/database/tables/user")
 const crypto = require("crypto");
 
-function view(req, res) {
+async function view(req, res) {
+    // Ensure user is connected
     if (require_connection(req, res))
         return;
 
+    // Ensure user has privileges
+    if (!await session_data(req).connected_user.can_create_repos())
+        return error_403(req, res, "Vous n'avez pas les droits pour créer un dépôt.");
+
     res.render('fileshare/fileshare', {
         title: 'Nouveau dépot',
-        user: req.session.user,
+        session_data: await session_data(req).client_data(),
+        public_data: await public_data(),
         force_open_create_repos: true,
     });
 }
 
 async function post_create_repos(req, res) {
-
+    // Ensure user is connected
     if (require_connection(req, res))
         return;
 
-    // Capture the input fields
+    // Ensure user has privileges
+    if (!await session_data(req).connected_user.can_edit_repos())
+        return error_403(req, res, "Vous n'avez pas les droits pour créer un dépôt.");
+
     let name = req.body.name;
-    let status = 'private'
+    let status = 'hidden'
     switch (req.body.type) {
-        case 'invisible':
+        case 'Invisible':
             status = 'hidden';
             break;
         case 'Privé':
@@ -35,23 +44,34 @@ async function post_create_repos(req, res) {
     }
 
     const access_key = crypto.randomBytes(16).toString("hex");
-    const new_repos = await Repos.insert(name, await Users.find(req.session.user.id), status, access_key)
-    req.session.user.my_repos.push(await new_repos.public_data())
+    await Repos.insert(name, await Users.find(req.session.user.id), status, access_key)
+
+    // Ensure connected user will refresh data
+    session_data(req).connected_user.mark_dirty();
+    if (status === 'public')
+        public_data().mark_dirty();
+
     res.redirect(`/fileshare/repos/${access_key}`);
 }
 
 async function post_delete_repos(req, res) {
-
     if (require_connection(req, res))
         return;
 
-    const repos = await Repos.find_access_key(req.params.repos)
+    // Ensure user has privileges
+    if (!await session_data(req).connected_user.can_edit_repos())
+        return error_403(req, res, "Vous n'avez pas les droits pour supprimer un dépôt.");
 
+    // Ensure the user who delete the repos is the owner
+    const repos = await Repos.find_access_key(req.params.repos)
     if ((await repos.get_owner()).get_id() !== req.session.user.id)
         return res.redirect(`/fileshare/`);
 
     await repos.delete();
-    req.session.user = await get_user_private_data(await Users.find(req.session.user.id));
+
+    // Ensure connected user will refresh data
+    events.on_delete_repos(repos);
+
     res.redirect(`/fileshare/`);
 }
